+++ snovault/types/user.py
@@ -61,7 +61,6 @@ class User(Item):
     item_type = 'user'
     schema = load_schema('snovault:schemas/user.json')
 
-    # TODO (C4-332): consolidate all acls into once place
     STATUS_ACL = {
         'current': ONLY_OWNER_VIEW_PROFILE_ACL,
         'deleted': DELETED_USER_ACL,
